Major refactoring of src folder
