+++ src/acceleration/Profiler.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "macros.h"
+#include "../macros.h"
 
 enum class Funcs {
   SceneConstruction,
