+++ src/acceleration/SafeProgressBar.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "macros.h"
+#include "../macros.h"
 
 void displayRenderProgress(float progress, int barWidth=70);
