+++ src/acceleration/SafeQueue.h
