+++ src/image/lodepng.cpp
