+++ src/vector/vector3d.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "macros.h"
+#include "../macros.h"
 
 class Vector3D {
 public:
